requirements.txt
requests
azure-identity
pandas
azure-storage-blob
Xlsxwriter
datetime
python-dotenv